Tighten up CI pipeline
The dry-release workflow now uses the new OIDC-based authentication flow
to authenticate with NPM, removing the need for access tokens. This is
to support NPM's move away from tokens in light of recent supply chain
attacks.

To support the new flow, the CI environment had to be upgraded to NPM
11.5.1, which requires Node 20.17 or later. However, we support Node all
back to 18.0.0, so the CI flow has been tweaked to run most of the
pipeline in latest LTS and install a different Node version just before
running the tests.

Note that although we support back to 18.0.0, our test suite runs only
on 18.14.0 or later. We won't be making a lot of changes till our next
major release, which will drop Node 18; so hopefully the coverage gap
won't cause big problems.

Finally, the 'npm publish --dry-run' command has been removed from CI.
Since it was running without release-it, the version number would always
conflict with the version on NPM and cause a failure.

"devEngines" has been set in package.json to help keep track of all this.